requirements.txt
streamlit
anthropic
elevenlabs
aci-sdk
pandas
python-dotenv
firebolt-sdk
